Refine the snow area analysis code to reduce repeated data processing and assign column default value as 'invalid'

+++ snow_cover_1.py
@@ -37,7 +37,9 @@ print 'step1: get snow date info'
 
 # user settings
 swe_ts_snow17 = 'DRGC2_we_snow17.ts'
-swe_ts_ueb = ''#'DRGC2_we_ueb.ts'
+swe_ts_ueb = 'DRGC2_uebW_local.ts'
+ueb_skip = 121
+snow17_skip = 136
 # snow_threshold = 0
 start_time = '2006-10-01'
 end_time = '2007-9-30'
@@ -48,14 +50,14 @@ concat = []
 col_name = []
 
 if swe_ts_snow17:
-    snow17_df = get_sim_dataframe(swe_ts_snow17, sim_skip=136, start_time=start_time, end_time=end_time)
+    snow17_df = get_sim_dataframe(swe_ts_snow17, sim_skip=snow17_skip, start_time=start_time, end_time=end_time)
     # snow17_snow_date = snow17_df[snow17_df > 0]
     # concat.append(snow17_snow_date)
     concat.append(snow17_df)
     col_name.append('snow17_swe')
 
 if swe_ts_ueb:
-    ueb_df = get_sim_dataframe(swe_ts_ueb, sim_skip=136, start_time=start_time, end_time=end_time)
+    ueb_df = get_sim_dataframe(swe_ts_ueb, sim_skip=ueb_skip , start_time=start_time, end_time=end_time)
     # ueb_snow_date = ueb_df[ueb_df > 0]
     # concat.append(ueb_snow_date)
     concat.append(ueb_df)
@@ -86,7 +88,7 @@ snow_date = pd.DataFrame.from_csv(model_snow_date_path, header=0)
 
 # unzip .gz file to assigned folder
 col_name = 'modis_snowdate_folder'
-snow_date[col_name] = np.nan
+snow_date[col_name] = 'invalid'
 for time in snow_date.index:
     year = str(time.year)
     day = '{:03d}'.format(time.dayofyear)
@@ -94,16 +96,18 @@ for time in snow_date.index:
     gz_name = 'modscag.snow_fraction_canadj.{}.{}.mosaic.tif.gz'.format(year, day)
     gz_path = os.path.join(modis_gz_folder, year, day, gz_name)
     modis_tif_path = os.path.join(modis_snowdate_folder, gz_name[:-3])
-
-    try:
-        with gzip.open(gz_path, 'rb') as inf:
-            with open(modis_tif_path, 'wb') as outf:
-                content = inf.read()
-                outf.write(content)
+    if not os.path.isfile(modis_tif_path):
+        try:
+            with gzip.open(gz_path, 'rb') as inf:
+                with open(modis_tif_path, 'wb') as outf:
+                    content = inf.read()
+                    outf.write(content)
+            snow_date[col_name].ix[time] = modis_tif_path
+        except Exception as e:
+            print 'failed to unzip modis data!!'
+            continue
+    else:
         snow_date[col_name].ix[time] = modis_tif_path
-    except Exception as e:
-        print 'failed to unzip modis data!!'
-        continue
 
 # step3 swe: unzip .gz file to assigned folder ####################################
 print 'step3: get swe data with snow'
@@ -111,7 +115,15 @@ print 'step3: get swe data with snow'
 # user settings
 swe_gz_folders = [
     r'/Projects/Tian_workspace/snow_analysis_test/WY2007/',  # snow17 swe xmrg .gz folder
-    # 'ueb_folder',  # ueb swe xmrg .gz folder
+    r'/Projects/Tian_workspace/rdhm_ueb_modeling/animas_2007_rec/SAC_out/uebW/',  # ueb swe xmrg .gz folder
+]
+grid_time_step = [
+    '06',  # snow17 swe output
+    '00'  # ueb swe output
+]
+var_name = [
+    'we',  # snow17 swe var name
+    'uebW',  # ueb swe var name
 ]
 
 swe_snowdate_folders = [os.path.join(result_folder, name) for name in ['snow17_snowdate_folder', 'ueb_snowdate_folder']]  # folder that stores swe tif with snow
@@ -125,20 +137,28 @@ for i in range(0, len(swe_gz_folders)):
         os.mkdir(swe_tif_folder)
 
     col_name = os.path.basename(swe_tif_folder)
-    snow_date[col_name] = ''
+    snow_date[col_name] = 'invalid'
 
     for time in snow_date.index:
-        swe_gz_name = 'we{}06z.gz'.format(time.strftime('%m%d%Y'))
+        swe_gz_name = '{}{}{}z.gz'.format(var_name[i], time.strftime('%m%d%Y'), grid_time_step[i])
         swe_gz_path = os.path.join(swe_gz_folder, swe_gz_name)
         swe_tif_path = os.path.join(swe_tif_folder, swe_gz_name[:-3])
-        try:
-            with gzip.open(swe_gz_path, 'rb') as inf:
-                with open(swe_tif_path, 'wb') as outf:
-                    content = inf.read()
-                    outf.write(content)
+        if not os.path.isfile(swe_tif_path):
+            try:
+                if os.path.isfile(swe_gz_path):
+                    with gzip.open(swe_gz_path, 'rb') as inf:
+                        with open(swe_tif_path, 'wb') as outf:
+                            content = inf.read()
+                            outf.write(content)
+                    snow_date[col_name].ix[time] = swe_tif_path
+            except Exception as e:
+                print 'failed to unzip model swe !'
+                print swe_gz_path
+                print os.path.isfile(swe_gz_path)
+                continue
+        else:
             snow_date[col_name].ix[time] = swe_tif_path
-        except Exception as e:
-            continue
+
     snow_date.to_csv(model_snow_date_path)
 
 print 'snow_cover_1: unzip files is done'
