+++ snow_cover_3.py
@@ -99,22 +99,26 @@ for swe_proj_folder in swe_proj_folders:
         # create dataframe column and folder
         proj_col_name = os.path.basename(swe_proj_folder)
         bin_col_name = proj_col_name.replace('proj', 'bin')
-        valid_date[bin_col_name] = ''
+        valid_date[bin_col_name] = 'invalid'
         swe_bin_folder = os.path.join(result_folder, bin_col_name)
         if not os.path.isdir(swe_bin_folder):
             os.mkdir(swe_bin_folder)
 
+        # create binary file if the file is not created
         for time in valid_date.index:
-            swe_proj_path =valid_date[proj_col_name].ix[time]
-            swe_bin_name = os.path.basename(swe_proj_path).replace('_proj', '_bin')
-            swe_bin_path = os.path.join(swe_bin_folder, swe_bin_name)
-            cmd_swe_bin = 'gdal_calc.py -A {} --calc="{}" --outfile={} --NoDataValue=-999 --type=Int16'.format(swe_proj_path, swe_func, swe_bin_path)
-
-            try:
-                subprocess.call(shlex.split(cmd_swe_bin))
-                valid_date[bin_col_name].ix[time] = swe_bin_path
-            except Exception as e:
-                continue
+            swe_proj_path = valid_date[proj_col_name].ix[time]
+            if os.path.isfile(swe_proj_path):
+                swe_bin_name = os.path.basename(swe_proj_path).replace('_proj', '_bin')
+                swe_bin_path = os.path.join(swe_bin_folder, swe_bin_name)
+                if not os.path.isfile(swe_bin_path):
+                    cmd_swe_bin = 'gdal_calc.py -A {} --calc="{}" --outfile={} --NoDataValue=-999 --type=Int16'.format(swe_proj_path, swe_func, swe_bin_path)
+                    try:
+                        subprocess.call(shlex.split(cmd_swe_bin))
+                        valid_date[bin_col_name].ix[time] = swe_bin_path
+                    except Exception as e:
+                        continue
+                else:
+                    valid_date[bin_col_name].ix[time] = swe_bin_path
 
         valid_date.to_csv(valid_date_path)
 
@@ -128,7 +132,7 @@ valid_date = pd.DataFrame.from_csv(valid_date_path, header=0)
 if os.path.isdir(modis_proj_folder):
     # create dataframe column and folder
     bin_col_name = 'modis_bin_folder'
-    valid_date[bin_col_name] = ''
+    valid_date[bin_col_name] = 'invalid'
     modis_bin_folder = os.path.join(result_folder, bin_col_name)
     if not os.path.isdir(modis_bin_folder):
         os.mkdir(modis_bin_folder)
@@ -137,51 +141,55 @@ if os.path.isdir(modis_proj_folder):
     modis_func = '((A>={})*(A<=100)+-999*(A==250))'.format(modis_threshold)  # when there is cloud, assign as no data values
     for time in valid_date.index:
         modis_proj_path = valid_date['modis_proj_folder'].ix[time]
-        modis_bin_name = os.path.basename(modis_proj_path).replace('_clip', '_bin')
-        modis_bin_path = os.path.join(modis_bin_folder, modis_bin_name)
-        cmd_modis_bin = 'gdal_calc.py -A {} --calc="{}" --outfile={} --NoDataValue=-999 --type=Int16'.format(modis_proj_path, modis_func, modis_bin_path)
+        if os.path.isfile(modis_proj_path):
+            modis_bin_name = os.path.basename(modis_proj_path).replace('_clip', '_bin')
+            modis_bin_path = os.path.join(modis_bin_folder, modis_bin_name)
 
-        try:
-            subprocess.call(shlex.split(cmd_modis_bin))
-            valid_date[bin_col_name].ix[time] = modis_bin_path
-        except Exception as e:
-            continue
-
-        # fill cloud cover area by interpolation
-        modis_proj_array = gdalnumeric.LoadFile(modis_proj_path)
-        cloud_cover_pixel = (modis_proj_array[0] == 250).sum()
-        if cloud_cover_pixel > 0:
-            from snow_cover_utility import create_bar_plot, array_to_raster
-            for i in range(10):
-                cmd_interp_bin = 'gdal_fillnodata.py -md 2 {} {}'.format(modis_bin_path, modis_bin_path)
-
+            if not os.path.isfile(modis_bin_path):
+                cmd_modis_bin = 'gdal_calc.py -A {} --calc="{}" --outfile={} --NoDataValue=-999 --type=Int16'.format(modis_proj_path, modis_func, modis_bin_path)
                 try:
-                    subprocess.call(shlex.split(cmd_interp_bin))
+                    subprocess.call(shlex.split(cmd_modis_bin))
+                    valid_date[bin_col_name].ix[time] = modis_bin_path
                 except Exception as e:
-                    valid_date[bin_col_name].ix[time] = ''
                     continue
 
-                modis_bin_array = gdalnumeric.LoadFile(modis_bin_path)
-                mask = np.ma.masked_where(modis_proj_array[1] == 0, modis_bin_array)  # remember to use the proj tif 2nd layer as the mask for binary tif creation
-                cloud_cover_pixel = (mask == -999).sum()
-                if cloud_cover_pixel == 0:
-                    break
+                # fill cloud cover area by interpolation
+                modis_proj_array = gdalnumeric.LoadFile(modis_proj_path)
+                cloud_cover_pixel = (modis_proj_array[0] == 250).sum()
+                if cloud_cover_pixel > 0:
+                    from snow_cover_utility import create_bar_plot, array_to_raster
+                    for i in range(10):
+                        cmd_interp_bin = 'gdal_fillnodata.py -md 2 {} {}'.format(modis_bin_path, modis_bin_path)
 
-            if cloud_cover_pixel > 0:
-                valid_date.drop(time)
-                print 'drop bin file at {}'.format(modis_bin_path)
+                        try:
+                            subprocess.call(shlex.split(cmd_interp_bin))
+                        except Exception as e:
+                            valid_date[bin_col_name].ix[time] = ''
+                            continue
+
+                        modis_bin_array = gdalnumeric.LoadFile(modis_bin_path)
+                        mask = np.ma.masked_where(modis_proj_array[1] == 0, modis_bin_array)  # remember to use the proj tif 2nd layer as the mask for binary tif creation
+                        cloud_cover_pixel = (mask == -999).sum()
+                        if cloud_cover_pixel == 0:
+                            break
+
+                    if cloud_cover_pixel > 0:
+                        valid_date.drop(time)
+                        print 'drop bin file at {}'.format(modis_bin_path)
+                    else:
+                        print modis_bin_path
+                        print os.path.isfile(modis_bin_path)
+                        os.remove(modis_bin_path)
+                        new_bin_array = np.where(modis_proj_array[1] != 0, modis_bin_array, -999)
+                        array_to_raster(output_path=modis_bin_path,
+                                        source_path=modis_proj_path,
+                                        array_data=new_bin_array)
+                        print 'cloud cover interpolation for {}'.format(modis_bin_path)
+
             else:
-                print modis_bin_path
-                print os.path.isfile(modis_bin_path)
-                os.remove(modis_bin_path)
-                new_bin_array = np.where(modis_proj_array[1] != 0, modis_bin_array, -999)
-                array_to_raster(output_path=modis_bin_path,
-                                source_path=modis_proj_path,
-                                array_data=new_bin_array)
-                print 'cloud cover interpolation for {}'.format(modis_bin_path)
+                valid_date[bin_col_name].ix[time] = modis_bin_path
 
     valid_date.to_csv(valid_date_path)
-    print 'modis binary file is done'
 
 else:
     print 'failed to create modis binary file'
